Add key rotation support
Issuer key can be rotated to minimise attack surface. This implies that
two keys might live at the same time, and that keys might have to be
cleaned after some time.

This commit adds support for keuy rotation. It does so by adding a daily
cron trigger. On the first of the month, a new key is added. Every day
which is not the first, keys are cleared so only the latest one
persists.
A new admin endpoint is also included /admin/clear to manually trigger
the clearing of keys.

Finally, the test suite is updated to reflect the underlying change in
key indexation, which now rely on truncated token id.