always use a prefix when interacting with R2
This will require a migration to deploy. To properly implement key
clearance we need to prefix all keys. In the case where a prefix is not
present, a `_global` prefix is used. This is because we don't want
listing of the regionless keys to return keys from all regions.

Luckily the migration is "additive" and not "destructive". We just need
to duplicate the current keys in R2 but with a `_global` prefix, and as
long as the deployment is finished before a rotation everything should
just work.